Fixed vc warning.
git-svn-id: https://irrlicht.svn.sourceforge.net/svnroot/irrlicht/trunk@1157 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CImageLoaderWAL.cpp
@@ -114,7 +114,7 @@ IImage* CImageLoaderWAL::loadImage(irr::io::IReadFile* file) const
 	// read image
 
 	const u32 imageSize = header.ImageHeight * header.ImageWidth;
-	if (file->getSize() < (imageSize + header.MipmapOffset[0]))
+	if (file->getSize() < (u32)(imageSize + header.MipmapOffset[0]))
 		return 0;
 
 	u8* data = new u8[imageSize];
